Move env variable test to integration test
